+++ src/pages/my_wallet/my_wallet.page.html
@@ -5,9 +5,21 @@
                 <ion-icon name="home"></ion-icon>
             </button>
         </ion-buttons>
+        <ion-buttons end>
+            <button ion-button (click)="stock_quotes_of_companies_page()" detail-none icon-only>
+                <ion-icon name="search"></ion-icon>
+            </button>
+        </ion-buttons>
         <ion-title text-center>My wallet</ion-title>
     </ion-navbar>
 </ion-header>
 <ion-content padding>
+
+    <h1 id="title" align="center">SELECTED COMPANIES STOCKS</h1>
+    <ul>
+      <li *ngFor="let stock of stocks">
+        {{ stock.symbol }} - {{ stock.companyName }} - {{ stock.latestPrice }} <button ion-button (click)="deleteStock(stock)">X</button>
+      </li>
+    </ul>
 </ion-content>
 <ion-nav [root]="rootPage" #content></ion-nav>
